Enum sqlite3_tokenizer::TokenKind

pub enum TokenKind {
    Semicolon,
    Explain,
    Query,
    Plan,
    Begin,
    Transaction,
    Deferred,
    Immediate,
    Exclusive,
    Commit,
    End,
    Rollback,
    Savepoint,
    Release,
    To,
    Table,
    Create,
    If,
    Not,
    Exists,
    Temp,
    LeftParen,
    RightParen,
    As,
    Without,
    Comma,
    Id,
    Indexed,
    Abort,
    Action,
    After,
    Analyze,
    Asc,
    Attach,
    Before,
    By,
    Cascade,
    Cast,
    ColumnKw,
    Conflict,
    Database,
    Desc,
    Detach,
    Each,
    Fail,
    For,
    Ignore,
    Initially,
    Instead,
    Like,
    Match,
    No,
    Key,
    Of,
    Offset,
    Pragma,
    Raise,
    Recursive,
    Replace,
    Restrict,
    Row,
    Trigger,
    Vacuum,
    View,
    Virtual,
    With,
    Reindex,
    Rename,
    CurrentTime,
    Any,
    Or,
    And,
    Is,
    Between,
    In,
    IsNull,
    NotNull,
    NotEqual,
    Equal,
    Greater,
    LessOrEqual,
    Less,
    GreaterOrEqual,
    Escape,
    BitAnd,
    BitOr,
    LeftShift,
    RightShift,
    Plus,
    Minus,
    Star,
    Slash,
    Remainder,
    Concat,
    Collate,
    BitNot,
    String,
    JoinKw,
    Constraint,
    Default,
    Null,
    Primary,
    Unique,
    Check,
    References,
    AutoIncrement,
    On,
    Insert,
    Delete,
    Update,
    Set,
    Deferrable,
    Foreign,
    Drop,
    Union,
    All,
    Except,
    Intersect,
    Select,
    Values,
    Distinct,
    Dot,
    From,
    Join,
    Using,
    Order,
    Group,
    Having,
    Limit,
    Where,
    Into,
    Integer,
    Float,
    Blob,
    Variable,
    Case,
    When,
    Then,
    Else,
    Index,
    Alter,
    Add,
    ToText,
    ToBlob,
    ToNumeric,
    ToInt,
    ToReal,
    IsNot,
    EndOfFile,
    Illegal,
    Space,
    UnclosedString,
    Function,
    Column,
    AggFunction,
    AggColumn,
    UnaryMinus,
    UnaryPlus,
    Register,
}

Describes the kind of token recognized

Many of these tokens are simply fixed words (e.g. SELECT) or punctuation.

Some of the more varied ones are:

Variable: A placeholder for a value in a prepared statement. ? or ?7 or @a or :b or #c

Space: Whitespace, but also comments.

Id: A user-defined identifier. my_column or [some table] or "that function" or `a trigger`.

String: A string literal. 'abcd' or 'planet''s moon'

Blob: A hexadecimal-encoded byte sequence literal. x'FEFF6564'

Variants

Semicolon

Explain

Query

Plan

Begin

Transaction

Deferred

Immediate

Exclusive

Commit

End

Rollback

Savepoint

Release

To

Table

Create

If

Not

Exists

Temp

LeftParen

RightParen

As

Without

Comma

Id

Indexed

Abort

Action

After

Analyze

Asc

Attach

Before

By

Cascade

Cast

ColumnKw

Conflict

Database

Desc

Detach

Each

Fail

For

Ignore

Initially

Instead

Like

Match

No

Key

Of

Offset

Pragma

Raise

Recursive

Replace

Restrict

Row

Trigger

Vacuum

View

Virtual

With

Reindex

Rename

CurrentTime

Any

Or

And

Is

Between

In

IsNull

NotNull

NotEqual

Equal

Greater

LessOrEqual

Less

GreaterOrEqual

Escape

BitAnd

BitOr

LeftShift

RightShift

Plus

Minus

Star

Slash

Remainder

Concat

Collate

BitNot

String

JoinKw

Constraint

Default

Null

Primary

Unique

Check

References

AutoIncrement

On

Insert

Delete

Update

Set

Deferrable

Foreign

Drop

Union

All

Except

Intersect

Select

Values

Distinct

Dot

From

Join

Using

Order

Group

Having

Limit

Where

Into

Integer

Float

Blob

Variable

Case

When

Then

Else

Index

Alter

Add

ToText

ToBlob

ToNumeric

ToInt

ToReal

IsNot

EndOfFile

Illegal

Space

UnclosedString

Function

Column

AggFunction

AggColumn

UnaryMinus

UnaryPlus

Register